chore: cardNickname page의 form label 위치 이동
- 추가: Button 컴포넌트를 Form 컴포넌트 밖으로 빼내는 과정에서 생긴 에러 해결
-- BrowserRouter 대신 createBrowserRouter 사용
