+++ react-payments-app/src/App.jsx
@@ -1,6 +1,6 @@
 import { RouterProvider } from 'react-router-dom';
 
-import router from './routers/router';
+import router from './router';
 
 const App = () => {
   return (
