+++ react-payments-app/src/components/common/Input/Form.jsx
@@ -1,9 +1,14 @@
-const Form = ({ handleSubmit, children, formErrorMessage }) => {
+import { Children } from 'react';
+import { Form as RouterForm } from 'react-router-dom';
+
+const Form = ({ id, handleSubmit, children, error }) => {
   return (
     <div id='card-form' className='w-90'>
-      <form onSubmit={handleSubmit}>{children}</form>
+      <form id={id} onSubmit={handleSubmit}>
+        {Children.toArray(children)}
+      </form>
       <div id='card-form-label' style={{ color: 'red' }}>
-        {formErrorMessage}
+        {error}
       </div>
     </div>
   );
