+++ react-payments-app/src/components/common/Input/Form.stories.jsx
@@ -18,7 +18,7 @@ const Template = (args) => <Form {...args} />;
 export const CardRegistrationForm = Template.bind({});
 CardRegistrationForm.args = {
   handleSubmit: () => alert('submitted'),
-  formErrorMessage: '닉네임은 10글자 이하로 작성해 주세요.',
+  error: '닉네임은 10글자 이하로 작성해 주세요.',
   children: (
     <CardProvider>
       <CardNumbersInput />
